refactor: renamed serialization strategy
