+++ simulator/src/serializers/json_record_serialization_strategy.py
@@ -4,6 +4,6 @@ from .record_serialization_template import RecordSerializationTemplate
 from ..models.raw_data.raw_data import RawData
 
 
-class JsonRecordSerialization(RecordSerializationTemplate):
+class JsonRecordSerializationStrategy(RecordSerializationTemplate):
     def serialize_value(self, raw_data: RawData) -> bytes:
         return json.dumps(raw_data.to_json()).encode('utf-8')
